refactor(fonts): extract font_loader.py with auto-discovery of latest release
