+++ tool/card_render.py
@@ -14,6 +14,8 @@ from io import BytesIO
 
 from PIL import Image, ImageDraw, ImageFont
 
+from . import font_loader
+
 # Module-level logger, can be overridden via set_logger()
 _module_logger = logging.getLogger(__name__)
 
@@ -22,6 +24,7 @@ def set_logger(logger: logging.Logger) -> None:
     """Set the logger used by card_render (call before init_fonts)."""
     global _module_logger
     _module_logger = logger
+    font_loader.set_logger(logger)
 
 
 # ─── 主题配色 ────────────────────────────────────────────────
@@ -84,13 +87,6 @@ THEME: dict[str, tuple] = THEME_DARK
 
 # ─── 字体管理 ───────────────────────────────────────────────
 
-_FONT_DOWNLOAD_URL = (
-    "https://mirrors.tuna.tsinghua.edu.cn/github-release/be5invis/Sarasa-Gothic/"
-    "Sarasa%20Gothic%2C%20Version%201.0.36/SarasaTermSlabSC-TTF-1.0.36.7z"
-)
-_DEFAULT_FONT_REGULAR = "SarasaTermSlabSC-Regular.ttf"
-_DEFAULT_FONT_BOLD = "SarasaTermSlabSC-Bold.ttf"
-
 # 运行时字体路径（由 init_fonts 设置）
 _font_regular_path: str = ""
 _font_bold_path: str = ""
@@ -98,252 +94,20 @@ _font_cache: dict[tuple[str, int], ImageFont.FreeTypeFont] = {}
 _fonts_ready = False
 
 
-def _find_fonts_in_dir(font_dir: str) -> tuple[str, str] | None:
-    """在目录中查找可用字体对 (regular, bold)，返回 (regular_path, bold_path) 或 None"""
-    if not os.path.isdir(font_dir):
-        return None
-    ttf_files = [f for f in os.listdir(font_dir) if f.lower().endswith(".ttf")]
-    if not ttf_files:
-        return None
-
-    regular = bold = None
-    for f in sorted(ttf_files):
-        fl = f.lower()
-        if "bold" in fl:
-            bold = bold or os.path.join(font_dir, f)
-        elif any(k in fl for k in ("regular", "normal", "medium")):
-            regular = regular or os.path.join(font_dir, f)
-
-    if not regular and not bold and ttf_files:
-        regular = bold = os.path.join(font_dir, ttf_files[0])
-    elif regular and not bold:
-        bold = regular
-    elif bold and not regular:
-        regular = bold
-
-    return (regular, bold) if regular and bold else None
-
-
-def _extract_7z(archive_path: str, output_dir: str, _log: logging.Logger) -> None:
-    """Extract a 7z archive. Tries system 7z binary first, then py7zr."""
-    import subprocess
-
-    for cmd in ("7z", "7za"):
-        try:
-            result = subprocess.run(
-                [cmd, "x", archive_path, f"-o{output_dir}", "-y"],
-                capture_output=True,
-                timeout=120,
-            )
-            if result.returncode == 0:
-                _log.info(f"使用 {cmd} 解压成功")
-                return
-        except (FileNotFoundError, subprocess.TimeoutExpired):
-            continue
-
-    _log.info("系统未安装 7z，尝试使用 py7zr 解压 ...")
-    try:
-        import py7zr
-    except ImportError as e:
-        raise RuntimeError(
-            "解压字体压缩包需要系统 7z 工具或 py7zr Python 包，"
-            "请安装之后重试：pip install py7zr"
-        ) from e
-
-    with py7zr.SevenZipFile(archive_path, "r") as z:
-        z.extractall(path=output_dir)
-
-
-def _download_fonts(font_dir: str) -> None:
-    """从清华镜像下载字体 7z 包并解压"""
-    import shutil
-    import tempfile
-    import threading
-    import urllib.request
-    from concurrent.futures import ThreadPoolExecutor, as_completed
-
-    _log = _module_logger
-    _NUM_THREADS = 4
-
-    os.makedirs(font_dir, exist_ok=True)
-    archive_path = os.path.join(font_dir, "_font_download.7z")
-
-    if not os.path.exists(archive_path):
-        _log.info("正在从清华镜像下载字体 ...")
-
-        supports_range = False
-        total_size = 0
-        try:
-            probe = urllib.request.Request(_FONT_DOWNLOAD_URL)
-            probe.add_header("Range", "bytes=0-0")
-            probe_resp = urllib.request.urlopen(probe, timeout=30)
-            if probe_resp.status == 206:
-                content_range = probe_resp.headers.get("Content-Range", "")
-                if "/" in content_range:
-                    total_size = int(content_range.split("/")[-1])
-                    supports_range = True
-            probe_resp.close()
-        except Exception:
-            pass
-
-        if supports_range and total_size > 1024 * 1024:
-            _log.info(
-                f"文件大小: {total_size / 1024 / 1024:.1f}MB, "
-                f"使用 {_NUM_THREADS} 线程下载"
-            )
-            chunk_ranges = []
-            chunk_sz = total_size // _NUM_THREADS
-            for i in range(_NUM_THREADS):
-                start = i * chunk_sz
-                end = (
-                    (total_size - 1)
-                    if i == _NUM_THREADS - 1
-                    else (start + chunk_sz - 1)
-                )
-                chunk_ranges.append((i, start, end))
-
-            downloaded_lock = threading.Lock()
-            downloaded_bytes = [0]
-            last_logged = [-1]
-
-            def _download_chunk(idx: int, byte_start: int, byte_end: int) -> str:
-                part_file = os.path.join(font_dir, f"_chunk_{idx}.part")
-                r = urllib.request.Request(_FONT_DOWNLOAD_URL)
-                r.add_header("Range", f"bytes={byte_start}-{byte_end}")
-                response = urllib.request.urlopen(r, timeout=180)
-                buf_size = 64 * 1024
-                with open(part_file, "wb") as f:
-                    while True:
-                        buf = response.read(buf_size)
-                        if not buf:
-                            break
-                        f.write(buf)
-                        with downloaded_lock:
-                            downloaded_bytes[0] += len(buf)
-                            pct = int(downloaded_bytes[0] / total_size * 100)
-                            if pct // 10 > last_logged[0] // 10:
-                                dl_mb = downloaded_bytes[0] / 1024 / 1024
-                                tot_mb = total_size / 1024 / 1024
-                                _log.info(
-                                    f"字体下载进度: {pct}% ({dl_mb:.1f}/{tot_mb:.1f}MB)"
-                                )
-                                last_logged[0] = pct
-                return part_file
-
-            part_files = [None] * _NUM_THREADS
-            try:
-                with ThreadPoolExecutor(max_workers=_NUM_THREADS) as pool:
-                    futures = {
-                        pool.submit(_download_chunk, idx, s, e): idx
-                        for idx, s, e in chunk_ranges
-                    }
-                    for future in as_completed(futures):
-                        idx = futures[future]
-                        part_files[idx] = future.result()
-
-                with open(archive_path, "wb") as out:
-                    for pf in part_files:
-                        with open(pf, "rb") as inp:
-                            shutil.copyfileobj(inp, out)
-                _log.info("字体下载完成，正在校验 ...")
-            finally:
-                for pf in part_files:
-                    if pf and os.path.exists(pf):
-                        os.remove(pf)
-        else:
-            if not supports_range:
-                _log.info("服务器不支持分段下载，使用单线程模式")
-            total_mb = total_size / 1024 / 1024 if total_size > 0 else 0
-            if total_mb:
-                _log.info(f"文件大小: {total_mb:.1f}MB")
-
-            resp = urllib.request.urlopen(
-                urllib.request.Request(_FONT_DOWNLOAD_URL), timeout=180
-            )
-            downloaded = 0
-            chunk_size = 64 * 1024
-            last_logged_pct = -1
-            part_path = archive_path + ".part"
-
-            with open(part_path, "wb") as f:
-                while True:
-                    chunk = resp.read(chunk_size)
-                    if not chunk:
-                        break
-                    f.write(chunk)
-                    downloaded += len(chunk)
-                    if total_size > 0:
-                        pct = int(min(downloaded / total_size, 1.0) * 100)
-                        if pct // 10 > last_logged_pct // 10:
-                            f.flush()
-                            os.fsync(f.fileno())
-                            _log.info(
-                                f"字体下载进度: {pct}% "
-                                f"({downloaded / 1024 / 1024:.1f}/{total_mb:.1f}MB)"
-                            )
-                            last_logged_pct = pct
-
-            os.rename(part_path, archive_path)
-            _log.info("字体下载完成，正在校验 ...")
-    else:
-        _log.info("检测到已下载的字体包，正在校验 ...")
-
-    _SEVEN_Z_MAGIC = b"7z\xbc\xaf\x27\x1c"
-    with open(archive_path, "rb") as f:
-        header = f.read(6)
-    if header != _SEVEN_Z_MAGIC:
-        _log.warning("下载的文件不是有效的 7z 压缩包，已删除，请重试")
-        os.remove(archive_path)
-        raise ValueError("Downloaded file is not a valid 7z archive")
-
-    try:
-        with tempfile.TemporaryDirectory() as tmp_dir:
-            _extract_7z(archive_path, tmp_dir, _log)
-
-            kept = 0
-            for root, _dirs, files in os.walk(tmp_dir):
-                for fname in files:
-                    if fname.lower().endswith(".ttf") and fname in (
-                        _DEFAULT_FONT_REGULAR,
-                        _DEFAULT_FONT_BOLD,
-                    ):
-                        shutil.copy2(os.path.join(root, fname), font_dir)
-                        kept += 1
-
-            _log.info(f"字体安装完成 ({kept} 个文件)")
-    finally:
-        if os.path.exists(archive_path):
-            os.remove(archive_path)
-
-
 def init_fonts(font_dir: str | None = None) -> bool:
-    """初始化字体。如果 font_dir 有字体就用，没有就自动下载。"""
+    """初始化字体。如果 font_dir 有字体就用，没有就自动下载（最新版本）。"""
     global _font_regular_path, _font_bold_path, _fonts_ready, _font_cache
 
     if font_dir is None:
         font_dir = os.path.join(os.path.dirname(__file__), "font")
 
-    found = _find_fonts_in_dir(font_dir)
-    if found:
-        _font_regular_path, _font_bold_path = found
-        _font_cache.clear()
-        _fonts_ready = True
-        return True
-
-    try:
-        _download_fonts(font_dir)
-    except Exception as e:
-        _module_logger.warning(f"字体下载失败: {e}")
+    paths = font_loader.init_fonts(font_dir)
+    if not paths:
         return False
-
-    found = _find_fonts_in_dir(font_dir)
-    if found:
-        _font_regular_path, _font_bold_path = found
-        _font_cache.clear()
-        _fonts_ready = True
-        return True
-
-    return False
+    _font_regular_path, _font_bold_path = paths
+    _font_cache.clear()
+    _fonts_ready = True
+    return True
 
 
 def _get_font(bold: bool = False, size: int = 18) -> ImageFont.FreeTypeFont:
